requirements.txt
PyPDF2==3.0.1
googletrans==4.0.0rc1
deep-translator==1.11.4
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
